% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_model.R
\name{boot_model}
\alias{boot_model}
\title{Bootstrap model fit and predictions}
\usage{
boot_model(
  imputed,
  outcome,
  seed = Sys.time(),
  family = "binomial",
  iter = 1000,
  s = c("lambda.min", "lambda.1se"),
  do_impute = TRUE,
  ...
)
}
\arguments{
\item{imputed}{An object of class \link{mids} containing the imputed data.}

\item{outcome}{A character string identifying the outcome measure.}

\item{seed}{The random seed to use.}

\item{family}{The \link{family} to be used for modelling the data. Assuming the use of logistic regression by default.}

\item{iter}{Number of bootstrap iterations to run.}

\item{s}{A character string indicating which value of \eqn{\lambda} should be used. Either \code{lambda.min}
for the optimal \eqn{\lambda} determined by cross-validation or \code{lambda.1se} for the optimal \eqn{\lambda + 1} standard error.}

\item{do_impute}{Logical indicating whether imputation of missing values should be performed as part of the bootstrap procedure.}

\item{...}{Further arguments passed to \link{performance}.}
}
\value{
A list with the results of calls to \link{fit_model} and \link{performance} for each bootstrap sample.
}
\description{
Bootstrap model fit and predictions
}
\details{
When \code{do_impute} is \code{TRUE} (the default), observations are sampled from the original dataset, including missing values.
These data as well as all parameters for the imputation of bootstrapped data are taken from \code{imputed} argument. When
\code{do_impute} is \code{FALSE} no new imputation takes place. Instead, resampling of the individuals in the original dataset
takes place in such a way that the observations from all imputed datasets are included for a selected individual.
}
\author{
Peter Humburg
}
