% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_imputed.R
\name{call_imputed}
\alias{call_imputed}
\title{Call binary outcomes from imputed data}
\usage{
call_imputed(x, outcome, threshold = 0.5)
}
\arguments{
\item{x}{A multiply imputed dataset}

\item{outcome}{Name of the outcome variable in \code{x}}

\item{threshold}{The threshold to be used for calling outcomes. Imputed outcomes with an average below this value will be called \emph{0}, all others \emph{1}.}
}
\value{
A numeric vector of \emph{0}s and \emph{1}s, indicating outcomes.
}
\description{
Call binary outcomes from imputed data
}
