% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Prepare data for use with glmnet}
\usage{
clean_data(data)
}
\arguments{
\item{data}{A \emph{data.frame}.}
}
\value{
A \emph{data.frame} of the same structure as \code{data} but with all
factors converted to integer vectors.
}
\description{
Prepare data for use with glmnet
}
\examples{
library(mice)
clean_data(nhanes2)
}
