% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_auc.R
\name{pool_auc}
\alias{pool_auc}
\title{Pool AUC estimates from multiply imputed data using Rubin's rules}
\usage{
pool_auc(roc, n, conf_level = 0.95)
}
\arguments{
\item{roc}{A list of objects of class \link{roc}}

\item{n}{Sample size}

\item{conf_level}{Size of confidence interval}
}
\value{
A numeric vector with elements \emph{estimate}, \emph{ci.lower}, and \emph{ci.upper} providing the pooled AUC
and its 95\% confidence interval respectively.
}
\description{
Pool AUC estimates from multiply imputed data using Rubin's rules
}
\author{
Peter Humburg
}
\keyword{internal}
