% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoslem.R
\name{hoslem}
\alias{hoslem}
\title{Compute Hosmer-Lemeshow Goodness of Fit statistic}
\usage{
hoslem(data, outcome, prediction, ...)
}
\arguments{
\item{data}{Multiply imputed dataset. Either an object of class \link{mids} or a data frame created with \link[mice:complete.mids]{complete()}.}

\item{outcome}{Name of the outcome variable.}

\item{prediction}{A vector of predicted values.}

\item{...}{Further arguments are passed on to \code{\link[generalhoslem:logitgof]{generalhoslem::logitgof()}}}
}
\value{
A list of objects of class \code{htest} with one entry per dataset.
}
\description{
Compute Hosmer-Lemeshow Goodness of Fit statistic
}
