% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance.binomial}
\alias{performance.binomial}
\title{Performance metrics for logistic regression models}
\usage{
\method{performance}{binomial}(
  x,
  data,
  outcome,
  metrics = c("roc", "auc", "specificity", "sensitivity", "accuracy", "precision",
    "brier", "r2", "hoslem"),
  model_fits,
  ...
)
}
\arguments{
\item{x}{The pooled logistic regression model to be evaluated.}

\item{data}{Data to be used in the evaluation.}

\item{outcome}{The name of the outcome measure. This should be the name of a variable in \code{data}.}

\item{metrics}{A character vector indicating the metrics to be computed}

\item{model_fits}{A list of fitted model objects. Should contain one model for each imputed dataset (required if 'r2' is requested).}
}
\description{
Performance metrics for logistic regression models
}
