% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance.gaussian}
\alias{performance.gaussian}
\title{Performance metrics for linear models}
\usage{
\method{performance}{gaussian}(
  x,
  data,
  outcome,
  metrics = c("r2", "adj.r2", "aic", "cor"),
  model_fits,
  ...
)
}
\arguments{
\item{x}{The model to be evaluated.}

\item{data}{Data to be used in the evaluation.}

\item{outcome}{The name of the outcome measure. This should be the name of a variable in \code{data}.}

\item{metrics}{A character vector indicating the metrics to be computed}

\item{model_fits}{A list of fitted model objects. Should contain one model for each imputed dataset.}

\item{...}{Further arguments to be passed to specific methods}
}
\value{
A list of requested performance metrics.
}
\description{
Performance metrics for linear models
}
