% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{Performance metrics for predictive models}
\usage{
performance(x, ...)
}
\arguments{
\item{x}{The model to be evaluated.}

\item{...}{Further arguments to be passed to specific methods}
}
\value{
A list of performance metrics.
}
\description{
Performance metrics for predictive models
}
\seealso{
\code{\link[=performance.binomial]{performance.binomial()}}, \code{\link[=performance.gaussian]{performance.gaussian()}}, \code{\link[=performance.cv]{performance.cv()}}
}
