% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cv}
\alias{plot.cv}
\title{Create plots to assess predictive performance following cross-validation}
\usage{
\method{plot}{cv}(x, ...)
}
\arguments{
\item{x}{An object of class \link{cv}.}

\item{...}{Further arguments are ignored.}
}
\description{
Create plots to assess predictive performance following cross-validation
}
