% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_auc.R
\name{pool_auc}
\alias{pool_auc}
\title{Pool AUC estimates from multiply imputed data using Rubin's rules}
\usage{
pool_auc(roc)
}
\arguments{
\item{roc}{A list of objects of class \link{roc}}
}
\value{
A list with elements \emph{estimate} and \emph{ci} providing the pooled AUC
and its 95\% confidence interval respectively.
}
\description{
Pool AUC estimates from multiply imputed data using Rubin's rules
}
\author{
Peter Humburg
}
