% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_hoslem.R
\name{pool_hoslem}
\alias{pool_hoslem}
\title{Pooling Hosmer-Lemeshow Goodness of Fit Statistic}
\usage{
pool_hoslem(pooled_model, data, outcome, prediction, iter = 1000, ...)
}
\arguments{
\item{pooled_model}{Model with pooled parameter estimates.}

\item{data}{Multiply imputed data. Either an object of class \link{mids} or a data frame as produced by \code{\link[mids:complete]{mids::complete()}}.}

\item{outcome}{Name of the column containing the outcome variable.}

\item{prediction}{Model predictions as a single vector.}

\item{iter}{Number of iterations for bootstrapping.}

\item{...}{Further arguments to be passed to \link{hoslem}.}
}
\value{
A vector with the pooled
}
\description{
Pooling Hosmer-Lemeshow Goodness of Fit Statistic
}
