% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_model.R
\name{pool_model}
\alias{pool_model}
\title{Performance metrics for predictive models}
\usage{
pool_model(model, data, ...)
}
\arguments{
\item{model}{List of fitted models}

\item{data}{Imputed data}

\item{...}{Further arguments to be passed to specific methods}
}
\value{
Pooled model object
}
\description{
Performance metrics for predictive models
}
\seealso{
\code{\link[=pool_model.binomial]{pool_model.binomial()}}, \code{\link[=pool_model.gaussian]{pool_model.gaussian()}}
}
