% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_predictions.R
\name{pool_predictions}
\alias{pool_predictions}
\title{Pool predictions}
\usage{
pool_predictions(pred)
}
\arguments{
\item{pred}{A matrix of predicted outcomes. One column per imputed dataset.}
}
\value{
A vector of pooled predictions.
}
\description{
Pool predictions
}
\author{
Peter Humburg
}
