% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_r2.R
\name{pool_r2.binomial}
\alias{pool_r2.binomial}
\title{Obtain pooled estimates of Nagelkerke's R-squared}
\usage{
\method{pool_r2}{binomial}(pooled_model, model_fits, data, iter = 1000)
}
\arguments{
\item{pooled_model}{The pooled model returned by \code{\link[=fit_model]{fit_model()}}.}

\item{model_fits}{List of model objects as returned by \code{\link[=fit_model]{fit_model()}}.}

\item{data}{Imputed data}

\item{iter}{Number of bootstrap iterations}

\item{...}{
  Arguments passed on to \code{\link[=pool_r2]{pool_r2}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
A numeric vector with the pooled estimate of $R^2$ and its variance.
}
\description{
Obtain pooled estimates of Nagelkerke's R-squared
}
