% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_r2.R
\name{pool_r2}
\alias{pool_r2}
\title{Pool estimates of R-squared}
\usage{
pool_r2(model, data, ...)
}
\arguments{
\item{data}{Imputed data}

\item{...}{Further arguments to be passed to specific methods}

\item{pooled_model}{The pooled model returned by \code{\link[=fit_model]{fit_model()}}.}

\item{model_fits}{List of model objects as returned by \code{\link[=fit_model]{fit_model()}}.}
}
\value{
A numeric vector with the pooled estimate of $R^2$ and its variance.
}
\description{
Pool estimates of R-squared
}
\seealso{
\code{\link[=pool_r2.binomial]{pool_r2.binomial()}}, \code{\link[=pool_r2.gaussian]{pool_r2.gaussian()}}
}
