% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_outcome.R
\name{predict_outcome}
\alias{predict_outcome}
\title{Predict outcomes for multiply imputed data}
\usage{
predict_outcome(fit, data, ...)
}
\arguments{
\item{fit}{A fitted model object to use for predictions.}

\item{data}{A multiply imputed dataset. Either of class \link{mids} or in \emph{long} format.}

\item{...}{Further arguments passed to \link{predict}.}
}
\value{
A matrix with one column for each imputed dataset.
}
\description{
Predict outcomes for multiply imputed data
}
\author{
Peter Humburg
}
