% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_data.R
\name{scale_data}
\alias{scale_data}
\title{Scale numeric variables}
\usage{
scale_data(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame of the same structure as \code{data} with all numeric variables centred and scaled.
}
\description{
Scale numeric variables
}
\author{
Peter Humburg
}
