% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_r2.R
\name{pool_r2}
\alias{pool_r2}
\title{Obtain pooled estimates of Nagelkerke's R-squared}
\usage{
pool_r2(model, data, iter = 1000)
}
\arguments{
\item{model}{List of model objects as returned by \link{fit_model}.}

\item{data}{Imputed dataset}

\item{iter}{Number of bootstrap iterations}
}
\value{
A numeric vector with the pooled estimate of $R^2$ and its variance.
}
\description{
Obtain pooled estimates of Nagelkerke's R-squared
}
