% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_r2.R
\name{pool_r2}
\alias{pool_r2}
\title{Obtain pooled estimates of R-squared}
\usage{
pool_r2(
  pooled_model,
  model_fits,
  data,
  iter = 1000,
  method = c("nagelkerke", "r2", "adj.r2")
)
}
\arguments{
\item{pooled_model}{The pooled model returned by \link{fit_model}.}

\item{model_fits}{List of model objects as returned by \link{fit_model}.}

\item{data}{Imputed dataset}

\item{iter}{Number of bootstrap iterations (only used for method="nagelkerke")}

\item{method}{Type of R^2 to compute and pool}
}
\value{
A numeric vector with the pooled estimate of $R^2$ and its variance.
}
\description{
Obtain pooled estimates of R-squared
}
