% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fit.R
\name{test_fit}
\alias{test_fit}
\title{Assessing model performance}
\usage{
test_fit(
  model,
  data,
  which = c("auc", "thresholds", "sp", "se", "coords"),
  bootstrap = FALSE,
  iter = 2000,
  level = 0.95
)
}
\arguments{
\item{model}{An object of class \link{glm}.}

\item{data}{Data to use in performance assessment.}

\item{which}{Character vector indicating which performance measures should be computed. One or more of \emph{auc}, \emph{thresholds}, \emph{sp}, \emph{se}, \emph{coords}.}

\item{bootstrap}{Logical indicating whether bootstrap samples should be generated from \code{data}.}

\item{iter}{Number of bootstrap iterations to use.}

\item{level}{Width of the confidence interval.}
}
\value{
A list of performance measures of the same length as \code{which}.
}
\description{
Assessing model performance
}
\details{
Confidence intervals may be calculated for the AUC (with \code{which="auc"}), thresholds derived from the ROC curve of the \code{model} on the provided \code{data} (\code{which="thresholds"}),
specificity (\code{which="sp"}), sensitivity (\code{which="se"}), and the ROC curve (\code{which="coords"}).
}
\author{
Peter Humburg
}
